Add lognormal distribution. Add Hierarchical. Not ready yet. Minor changes in run_test.sh

+++ tests/REG-001/run_test.sh
@@ -12,6 +12,7 @@
 #     and rc = 0.  
 ###############################################################################
 
+
 ###### Auxiliar Functions and Variables #########
 
 source ../functions.sh
@@ -41,14 +42,29 @@ for dir in ./*
 do
   logEcho "-------------------------------------"
   logEcho " Entering Tutorial: $dir"
+
+  if [ ! -d "$dir/python" ]; then
+    echo "  + No folder named 'python' found inside $dir"
+    continue
+  fi
+
   pushd $dir/python >> $logFile 2>&1
   
+
   log "[Korali] Removing any old result files..."
   rm -rf _korali_results >> $logFile 2>&1
   check_result
   
-  for file in *.py
+  for file in run-*.py
   do
+    
+    # In case there are no files matching the pattern 'run-*.py', break
+    if [ ! -f "$file" ]; then
+      echo "  + No run*-.py file found!"
+      continue
+    fi
+    
+
     logEcho "  + Running File: ${file%.*}"
 
     log "[Korali] Adding Random Seed..."
